config: first pass at a config module.
This module will store immutable config data. The non-test interface to
this package presents a single Config::get() method which retrieves a
config that has been previously set by Config::set(). That is, this
module effectively implements a singleton config. I implemented things
this way because harpoxide will eventually read a config from disk at
startup & use it for the entire runtime of the server.